+++ src/JicoDotNet.Inventory.Core/Entities/IShipmentDetail.cs
@@ -1,6 +1,8 @@
-﻿namespace JicoDotNet.Inventory.Core.Entities
+﻿using JicoDotNet.Inventory.Core.Entities.Inner;
+
+namespace JicoDotNet.Inventory.Core.Entities
 {
-    public interface IShipmentDetail : IProductAttribute
+    public interface IShipmentDetail : IProductAttribute, IGenericDescription
     {
         long ShipmentDetailId { get; set; }
         long ShipmentId { get; set; }
@@ -9,6 +11,5 @@
         long ProductId { get; set; }
         decimal ShippedQuantity { get; set; }
         long StockDetailId { get; set; }
-        string Description { get; set; }
     }
 }