+++ src/JicoDotNet.Inventory.Core/Entities/Inner/IGenericDescription.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace JicoDotNet.Inventory.Core.Entities.Inner
+{
+    public interface IGenericDescription
+    {
+        string Description { get; set; }
+    }
+}